+++ crates/modular_core/src/dsp/oscillators/sine.rs
@@ -50,7 +50,7 @@ impl SineOscillator {
     fn update(&mut self, sample_rate: f32) {
         let num_channels = self.channel_count();
 
-        self.outputs.sample.set_channels(num_channels as u8);
+        self.outputs.sample.set_channels(num_channels);
 
         for ch in 0..num_channels {
             let state = &mut self.channels[ch];
