+++ crates/modular_core/src/dsp/utilities/sample_and_hold.rs
@@ -48,7 +48,7 @@ impl Default for SampleAndHold {
 impl SampleAndHold {
     pub fn update(&mut self, _sample_rate: f32) {
         let num_channels = self.channel_count();
-        self.outputs.sample.set_channels(num_channels as u8);
+        self.outputs.sample.set_channels(num_channels);
 
         for ch in 0..num_channels {
             let state = &mut self.channels[ch];
